Inline methods.
